Trade components created , preferences progress
