+++ smallchange/src/app/app.component.html
@@ -2,5 +2,5 @@
 <app-nav-bar></app-nav-bar>
 
 <!-- <app-login-page></app-login-page> -->
-<!-- <app-trade-page></app-trade-page> -->
-<app-preferences-page></app-preferences-page>
+<app-trade-page></app-trade-page>
+<!-- <app-preferences-page></app-preferences-page> -->
