+++ smallchange/src/app/organisms/edit-preference/edit-preference.component.ts
@@ -7,67 +7,19 @@ import { Component, OnInit } from '@angular/core';
 })
 export class EditPreferenceComponent implements OnInit {
 
-  enter: boolean = false;
-  edit: boolean = false;
-  flag: boolean = false;
-  agreed: boolean = false;
-  setEditTrue: boolean = false;
-  preferencesSet: boolean = true;
-  riskTolerance = '';
-  incomeCategory = '';
-  investmentLength = ''
-  investmentPurpose ='';
+  setEditTrue: boolean | undefined;
+  preferencesSet: boolean | undefined;
+  riskTolerance = 'HIGH';
+  incomeCategory = '2000';
+  investmentLength = '0-5 YEARS'
+  investmentPurpose ='RANDOM';
 
 
   constructor() { }
 
   ngOnInit(): void {
-  }
-
-  enterPreferences(): void {
-    console.log(this.enter);
-    
-    this.edit=false;
-
-    if (this.enter==true)
-    this.enter=false;
-    else{
-    this.enter = true;
-      if(!this.preferencesSet && this.flag )
-      this.preferencesSet = true;
-
-  }
-
-
-  }
-
-  setflag():void {
-    this.flag = true;
-    this.enter = false;
+    this.setEditTrue = false;
     this.preferencesSet = true;
-  }
-
-  editPreferences(): void {
-    this.enter = false;
-    if (this.edit==true)
-      this.edit = false;
-      
-    else
-      this.edit = true;
-    // if(this.preferencesSet==false)
-  
-  }
-  setAgreement():void{
-    if(this.agreed==false)
-    {
-      this.agreed = true;
-      // console.log(this.agreed);
-      
-    }
-    else{
-      this.agreed = false;
-      // console.log(this.agreed);
-    }
   }
 
   setEdit():void{
@@ -75,7 +27,14 @@ export class EditPreferenceComponent implements OnInit {
   }
 
   finishEdit():void{
-    this.setEditTrue = false;
+
   }
 
-}
+  }
+
+
+
+ 
+
+
+
