+++ smallchange/src/app/organisms/enter-preference/enter-preference.component.html
@@ -61,3 +61,5 @@
         <button mat-raised-button color="primary" (click)="setPreference()" [disabled]="!agreed">Submit</button>
     </div>
 </div>
+
+
