+++ smallchange/src/app/organisms/trade-buy/trade-buy.component.html
@@ -1,0 +1,3 @@
+<div class="trade-buy">
+<p>trade buy works</p>
+</div>
