+++ smallchange/src/app/organisms/trade-sell/trade-sell.component.html
@@ -1,0 +1,3 @@
+<div class="trade-sell">
+    <p>trade-sell works!</p>
+</div>
