+++ smallchange/src/app/organisms/trade-sell/trade-sell.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-trade-sell',
+  templateUrl: './trade-sell.component.html',
+  styleUrls: ['./trade-sell.component.css']
+})
+export class TradeSellComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}
