+++ smallchange/src/app/pages/preferences-page/preferences-page.component.html
@@ -1,18 +1,14 @@
 
 <div class=  "preference-page">
-    <div class="preferences-page-buttons">
-    <!-- <app-trade-button class="trade-button" (click)="enterPreferences()">Enter Preferences</app-trade-button>
-    <app-trade-button class="trade-button" (click)="editPreferences()">Edit Preferences</app-trade-button> -->
-</div>
 
 
 <div class="tablist">
     <mat-tab-group mat-stretch-tabs>
-    <mat-tab label="First"  >
+    <mat-tab label="Enter Preferences"  >
         
   <app-enter-preference></app-enter-preference>
  </mat-tab>
-    <mat-tab label="Second"  > 
+    <mat-tab label="Edit Preferences"> 
         <app-edit-preference></app-edit-preference>
     </mat-tab>
 </mat-tab-group>
