+++ smallchange/src/app/pages/trade-page/trade-page.component.css
@@ -1,3 +1,10 @@
 .trade-page-buttons{
     text-align-last: center;
 }
+
+.trade-page{
+    background-color: lightgray;
+}
+.mat-tab-group{
+align-items:center;
+}
