+++ smallchange/src/app/pages/trade-page/trade-page.component.html
@@ -1,34 +1,12 @@
-
-<div class="trade-page-buttons">
-    <app-trade-button class="trade-button" (click)="performBuy()">Execute Trade Buy</app-trade-button>
-    <app-trade-button class="trade-button" (click)="performSell()">Execute Trade Sell</app-trade-button>
+<div class="trade-page">
+<div class="tablist">
+  <mat-tab-group mat-stretch-tabs>
+  <mat-tab label="Trade Buy"  >
+      <app-trade-buy></app-trade-buy>
+</mat-tab>
+  <mat-tab label="Trade Sell"  > 
+      <app-trade-sell></app-trade-sell>
+  </mat-tab>
+</mat-tab-group>
 </div>
-
-<div *ngIf="buy">
-    Buy happened
 </div>
-
-<div *ngIf="sell">
-    Sell happened
-</div>
-
-<!-- <mat-carousel
-  timings="250ms ease-in"
-  [autoplay]="true"
-  color="accent"
-  maxWidth="50%"
-  [loop]="true"
-  [hideArrows]="false"
-  [hideIndicators]="false"
-  [useKeyboard]="true"
-  [useMouseWheel]="false"
-  orientation="ltr"
->
-  <mat-carousel-slide
-    #matCarouselSlide
-    *ngFor="let slide of slides; let i = index"
-    [image]="slide.image"
-    overlayColor="#00000040"
-    [hideOverlay]="false"
-  ></mat-carousel-slide>
-</mat-carousel> -->
